+++ R/anything.R
@@ -7,9 +7,12 @@
 #' (except line breaks) 0 or more times.
 #'
 #' @param .data Expression to append, typically pulled from the pipe \code{ \%>\% }
+#' @param mode Matching mode (\code{greedy} (default) or\code{lazy}). \code{Lazy} matching stops after the first match, \code{greedy} continues
+#' searching until end of the string and then back-tracks to the last match.
 #'
 #' @examples
 #' anything()
+#' anything(mode = "lazy")
 #'
 #' x <- start_of_line() %>%
 #'   anything() %>%
@@ -23,7 +26,13 @@
 #' Dot: \url{https://www.regular-expressions.info/dot.html}
 #'
 #' Star Quantifier: \url{https://www.regular-expressions.info/repeat.html}
+#'
+#' Greedy and Lazy Quantifiers: \url{https://www.regular-expressions.info/repeat.html#greedy}
 #' @export
-anything <- function(.data = NULL) {
-  paste0(.data, "(?:.*)")
+anything <- function(.data = NULL, mode = "greedy") {
+  switch(mode,
+    greedy = paste0(.data, "(?:.*)"),
+    lazy = paste0(.data, "(?:.*?)"),
+    stop("Please, provide valid 'mode' argument")
+  )
 }