+++ R/something_but.R
@@ -7,6 +7,8 @@
 #'
 #' @param .data Expression to append, typically pulled from the pipe \code{ \%>\% }
 #' @param value Expression to optionally match
+#' @param mode Matching mode (\code{greedy} (default) or\code{lazy}). \code{Lazy} matching stops after the first match, \code{greedy} continues
+#' searching until end of the string and then back-tracks to the last match.
 #'
 #' @examples
 #' something_but(value = "abc")
@@ -19,7 +21,13 @@
 #'
 #' @references
 #' Metacharacters: \url{https://www.regular-expressions.info/characters.html#special}
+#'
+#' Greedy and Lazy Quantifiers: \url{https://www.regular-expressions.info/repeat.html#greedy}
 #' @export
-something_but <- function(.data = NULL, value) {
-  paste0(.data, "(?:[^", sanitize(value), "]+)")
+something_but <- function(.data = NULL, value, mode="greedy") {
+  switch(mode,
+    greedy = paste0(.data, "(?:[^", sanitize(value), "]+)"),
+    lazy = paste0(.data, "(?:[^", sanitize(value), "]+?)"),
+    stop("Please, provide valid 'mode' argument")
+  )
 }