+++ NEWS.md
@@ -2,3 +2,4 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Deprecated `then()` and `any` in favor of `any_of()` and `find()`. 
+* Added lazy mode for `anything()`, `anything_but()`, `something()`, `something_but()`.